release

+++ src/sample/Model.java
@@ -1,7 +1,7 @@
 package sample;
 
 public class Model {
-    public float Calculate(long number1 ,long number2, String operator) {
+    public float Calculate(float number1 ,float number2, String operator) {
         switch (operator) {
             case "+":
             return number1 + number2;
